upload button abstraction
